sint_conv_layer compiling but gets wrong result in simulation.
